refactor(export,docs): demesher follow-ups — prune module split, duplicate-id guard, docs
- extract the reverse-reference mark-and-sweep into demesh-prune.ts
  (demesh-writer.ts 461 -> ~316 lines, both under the ~400 guideline)
- applySimplifiedGeometry replaces a repeated express id once and skips
  duplicates ('duplicate-id') instead of orphaning the first overlay chain
- document the complete-entityIndex.byId requirement (deferPropertyAtomIndex
  stores) and add a CLI docs section for simplify incl. the triangle-count
  vs bytes expectation
